requirements.txt
pytest
langchain-ollama
langchain-community
python-dotenv
requests
yfinance
pipdeptree
langchain
google-search-results
langchain-openai
textblob
selenium
webdriver-manager
undetected-chromedriver
cloudscraper







